+++ gl_additions/addons/GLA/CfgVehicles.hpp
@@ -1,0 +1,3 @@
+class CfgVehicles {
+
+};
